Removed dead code, out of date TODOs and some warns
