+++ conf/application.conf
@@ -209,11 +209,6 @@ mongodb {
   uri = "mongodb://localhost:27017/vat-registration"
 }
 
-api {
-  vatRestartURL = "http://localhost:9895/register-for-vat/before-you-register-for-vat"
-  vatCancelURL = "http://localhost:9895/internal/:regID/delete"
-}
-
 include "thresholds.conf"
 
 microservice {
@@ -241,11 +236,6 @@ microservice {
       url = "http://localhost:9896"
       environment = "local"
       authorization-token = "Fake Token"
-    }
-
-    business-registration {
-      host = localhost
-      port = 9660
     }
 
     incorporation-information {
