--- test/builders/SessionBuilder.scala
@@ -1,47 +1,0 @@
-/*
- * Copyright 2022 HM Revenue & Customs
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package builders
-
-import java.util.UUID
-
-import play.api.mvc.{AnyContentAsEmpty, AnyContentAsFormUrlEncoded}
-import play.api.test.FakeRequest
-import uk.gov.hmrc.http.SessionKeys
-
-object SessionBuilder extends SessionBuilder {}
-
-trait SessionBuilder {
-
-  val userIdKey: String = "userId"
-  val tokenKey: String = "token"
-
-  def updateRequestFormWithSession(fakeRequest: FakeRequest[AnyContentAsFormUrlEncoded], userId: String): FakeRequest[AnyContentAsFormUrlEncoded] = {
-    val sessionId = s"session-${UUID.randomUUID}"
-    fakeRequest.withSession(
-      SessionKeys.sessionId -> sessionId,
-      tokenKey -> "RANDOMTOKEN",
-      userIdKey -> userId)
-  }
-
-  def buildRequestWithSession(userId: String): FakeRequest[AnyContentAsEmpty.type] = {
-    val sessionId = s"session-${UUID.randomUUID}"
-    FakeRequest().withSession(
-      SessionKeys.sessionId -> sessionId,
-      tokenKey -> "RANDOMTOKEN",
-      userIdKey -> userId)
-  }
-}
